+++ tests/integration/Utility/AppTest.php
@@ -71,8 +71,7 @@ class AppTest extends TestCase
     {
         $app = new App(new Streams(null, 'php://output'));
         $this->expectOutputRegex(
-            "{^pipelines version (@\.@\.@|[a-f0-9]{7}|\d+\.\d+\.\d+)(-\d+-g[a-f0-9]{7})?\+?\n" .
-            "info: --file overrides non-default --basename\n}"
+            "{^info: --file overrides non-default --basename\n}"
         );
         $args = array(
             'pipelines-test',